# api/docs/swagger.yaml
definitions:
  accommodation.CreateHouseRes:
    properties:
      address:
        type: string
      bathrooms:
        type: integer
      bedrooms:
        type: integer
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      image_url:
        items:
          type: string
        type: array
      latitude:
        type: number
      lease_duration:
        type: integer
      lease_terms:
        type: string
      listing_status:
        type: string
      longitude:
        type: number
      owner_id:
        type: string
      price:
        type: number
      property_type:
        type: string
      roommate_count:
        type: integer
      square_footage:
        type: number
      top_status:
        type: boolean
      updated_at:
        type: string
    type: object
  accommodation.DeleteHouseRes:
    properties:
      message:
        type: string
    type: object
  accommodation.GetAllHouseRes:
    properties:
      house:
        items:
          $ref: '#/definitions/accommodation.GetAllHouses'
        type: array
    type: object
  accommodation.GetAllHouses:
    properties:
      address:
        type: string
      bathrooms:
        type: integer
      bedrooms:
        type: integer
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      image_url:
        items:
          type: string
        type: array
      latitude:
        type: number
      lease_duration:
        type: integer
      lease_terms:
        type: string
      listing_status:
        type: string
      longitude:
        type: number
      owner_id:
        type: string
      price:
        type: number
      property_type:
        type: string
      roommate_count:
        type: integer
      square_footage:
        type: number
      top_status:
        type: boolean
      updated_at:
        type: string
    type: object
  accommodation.GetByIdHouseRes:
    properties:
      address:
        type: string
      bathrooms:
        type: integer
      bedrooms:
        type: integer
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      image_url:
        items:
          type: string
        type: array
      latitude:
        type: number
      lease_duration:
        type: integer
      lease_terms:
        type: string
      listing_status:
        type: string
      longitude:
        type: number
      owner_id:
        type: string
      price:
        type: number
      property_type:
        type: string
      roommate_count:
        type: integer
      square_footage:
        type: number
      top_status:
        type: boolean
      updated_at:
        type: string
    type: object
  accommodation.UpdateHouseRes:
    properties:
      message:
        type: string
    type: object
  favorites.CreateFavoritesReq:
    properties:
      property_id:
        type: string
      user_id:
        type: string
    type: object
  favorites.CreateFavoritesRes:
    properties:
      created_at:
        type: string
      id:
        type: string
      property_id:
        type: string
      user_id:
        type: string
    type: object
  favorites.DeleteFavoritesRes:
    properties:
      message:
        type: string
    type: object
  favorites.Favorite:
    properties:
      created_at:
        type: string
      id:
        type: string
      property_id:
        type: string
      user_id:
        type: string
    type: object
  favorites.GetAllFavoritesRes:
    properties:
      favorites:
        items:
          $ref: '#/definitions/favorites.Favorite'
        type: array
    type: object
  favorites.GetByIdFavoritesRes:
    properties:
      created_at:
        type: string
      id:
        type: string
      property_id:
        type: string
      user_id:
        type: string
    type: object
...
  models.CreateProperties:
    properties:
      address:
        type: string
      bathrooms:
        type: integer
      bedrooms:
        type: integer
      description:
        type: string
      image_url:
        items:
          type: string
        type: array
      latitude:
        type: number
      lease_duration:
        type: integer
      lease_terms:
        type: string
      listing_status:
        type: string
      longitude:
        type: number
      price:
        type: number
      property_type:
        type: string
      roommate_count:
        type: integer
      square_footage:
        type: number
      top_status:
        type: boolean
    type: object
...
  models.UpdateProperties:
    properties:
      address:
        type: string
      bathrooms:
        type: integer
      bedrooms:
        type: integer
      description:
        type: string
      id:
        type: string
      image_url:
        items:
          type: string
        type: array
      latitude:
        type: number
      lease_duration:
        type: integer
      lease_terms:
        type: string
      listing_status:
        type: string
      longitude:
        type: number
      price:
        type: number
      property_type:
        type: string
      roommate_count:
        type: integer
      square_footage:
        type: number
      top_status:
        type: boolean
    type: object
...
  notification.CreateNotificationRequest:
    properties:
      content:
        type: string
      is_read:
        type: boolean
      user_id:
        type: string
    type: object
  notification.CreateNotificationResponse:
    properties:
      notification:
        $ref: '#/definitions/notification.Notification'
    type: object
  notification.GetNotificationResponse:
    properties:
      notification:
        $ref: '#/definitions/notification.Notification'
    type: object
  notification.Notification:
    properties:
      content:
        type: string
      created_at:
        type: string
      id:
        type: string
      is_read:
        type: boolean
      user_id:
        type: string
    type: object
...
info:
  contact: {}
  description: This is a sample server for Api-gateway Service
  title: Api Gateway
  version: "1.0"
paths:
  /favorites/create:
    post:
      consumes:
      - application/json
      description: Add a new property to user's favorites
      parameters:
      - description: Create Favorite Request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/favorites.CreateFavoritesReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/favorites.CreateFavoritesRes'
        "400":
          description: Bad Request
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Create a new favorite
      tags:
      - favorites
  /favorites/delete/{id}:
    delete:
      consumes:
      - application/json
      description: Remove a property from user's favorites
      parameters:
      - description: Favorite ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/favorites.DeleteFavoritesRes'
        "400":
          description: Bad Request
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Delete a favorite
      tags:
      - favorites
  /favorites/getall:
    get:
      consumes:
      - application/json
      description: Retrieve a list of all user's favorite properties
      parameters:
      - description: Limit
        in: query
        name: limit
        required: true
        type: integer
      - description: Page
        in: query
        name: page
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/favorites.GetAllFavoritesRes'
        "400":
          description: Bad Request
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Get all favorites
      tags:
      - favorites
  /favorites/getbyid/{id}:
    get:
      consumes:
      - application/json
      description: Retrieve a specific favorite by its ID
      parameters:
      - description: Favorite ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/favorites.GetByIdFavoritesRes'
        "400":
          description: Bad Request
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Get a favorite by ID
      tags:
      - favorites
  /notification/create:
    post:
      consumes:
      - application/json
      description: Create a new notification for a user
      parameters:
      - description: Create Notification
        in: body
        name: notification
        required: true
        schema:
          $ref: '#/definitions/notification.CreateNotificationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/notification.CreateNotificationResponse'
        "400":
          description: Bad Request
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Create a new notification
      tags:
      - notifications
  /notification/get/{id}:
    get:
      consumes:
      - application/json
      description: Get notification details by ID
      parameters:
      - description: Notification ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/notification.GetNotificationResponse'
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Get a notification by ID
      tags:
      - notifications
...
  /properties/propertiescreate:
    post:
      consumes:
      - application/json
      description: Create a new house with specified properties
      parameters:
      - description: Create House Request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.CreateProperties'
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            $ref: '#/definitions/accommodation.CreateHouseRes'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Create a new house
      tags:
      - houses
  /properties/propertiesdelete/{properties_id}:
    delete:
      consumes:
      - application/json
      description: Delete a house by ID
      parameters:
      - description: Property ID
        in: path
        name: properties_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            $ref: '#/definitions/accommodation.DeleteHouseRes'
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Delete a house
      tags:
      - houses
  /properties/propertiesgetall/{limit}/{page}:
    get:
      consumes:
      - application/json
      description: Retrieve a list of all houses with pagination
      parameters:
      - description: Limit
        in: path
        name: limit
        required: true
        type: integer
      - description: Page
        in: path
        name: page
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            $ref: '#/definitions/accommodation.GetAllHouseRes'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Get all houses
      tags:
      - houses
  /properties/propertiesgetbyid/{properties_id}:
    get:
      consumes:
      - application/json
      description: Retrieve house information by ID
      parameters:
      - description: Property ID
        in: path
        name: properties_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            $ref: '#/definitions/accommodation.GetByIdHouseRes'
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Get house by ID
      tags:
      - houses
  /properties/propertiesupdate:
    put:
      consumes:
      - application/json
      description: Update an existing house with new details
      parameters:
      - description: Update House Request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.UpdateProperties'
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            $ref: '#/definitions/accommodation.UpdateHouseRes'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Update an existing house
      tags:
      - houses
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"